+++ include/aikido/planner.hpp
@@ -13,3 +13,7 @@
 #include "planner/ompl/dart.hpp"
 #include "planner/parabolic/ParabolicSmoother.hpp"
 #include "planner/parabolic/ParabolicTimer.hpp"
+#include "planner/postprocessor/postprocess.hpp"
+#include "planner/postprocessor/RetimeTrajectoryPostProcessor.hpp"
+#include "planner/postprocessor/SmoothTrajectoryPostProcessor.hpp"
+#include "planner/postprocessor/TrajectoryPostProcessor.hpp"
